veterinaria/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.32.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
